Remove pre command hook
The only purpose of `adob--pre-command-hook' was to record the
`current-buffer' before the command was executed.  This allowed
`adob--post-command-hook' to check whether the buffer changed.

A much simpler approach is to record buffer whenever it is
undimmed.  This gets rid of the pre command hook and the whole
logic is now in `adob--post-command-hook'.