Drop turn-{on,off}-* functions.
`turn-on-auto-dim-other-buffers' and `turn-off-auto-dim-other-buffers'
functions were not really useful.  For example, they did not set the
`auto-dim-other-buffers-mode' variable so they could not be used
instead of calling the minor mode.  Furthermore, since [cfab476: Move
hooks adding/removing to separate function], they were pretty short
thus separating them did not provide much benefits as far as
readability is concerned.  Because of that, the functions are now
removed, and their code put directly into the minor mode function.