Move minibuffer check to the beginning of ‘adob--buffer-list-update-hook’
With ‘auto-dim-other-buffers-dim-on-switch-to-minibuffer’ option disabled,
if user switches from buffer α to minibuffer and then to buffer β (e.g.
by invoking ‘other-window’), buffer α will remain undimmed even though
β becomes the current one.

This is because ‘adob--last-buffer’ is updated to point to minibuffer
even though α remains undimmed.

Move the minibuffer check to the top of ‘adob--buffer-list-update-hook’
function so that switching to minibuffer becomes a noöp, i.e. the code
behaves as if the old buffer was still active.

Fixes: https://github.com/mina86/auto-dim-other-buffers.el/issues/14
Fixes: https://github.com/mina86/auto-dim-other-buffers.el/pull/15