Make sure current buffer is not dimmed when the mode is enabled
Procedure for enabling the mode was to dim all buffers including the
current one.  This left user in an unexpected state which normally
should not occur.  (Switching buffers or windows would rectify it but
that situation was nonetheless a bug).

Fix the issue by making sure current buffer is not dimmed during
mode’s initialisation.