Dim individual windows rather than buffers
It has been long standing bug where ‘auto-dim-other-buffers’ would not
dim windows if they displayed selected buffer.  I.e. if a buffer was
shown in multiple windows, and one of those windows was selected, all
of them would be highlighted.

This is now fixed so long as the mode is used in Emacs 27.1 or newer.
Even if a buffer is displayed in multiple windows, only window with
focus will be undimmed.

This is thanks to ‘:filtered (:window …)’ face predicate introduced in
27.1 which allows to activate or deactivated based on window
properties.

Fixes: https://github.com/mina86/auto-dim-other-buffers.el/issues/16
Fixes: https://github.com/mina86/auto-dim-other-buffers.el/issues/19