Unconditionally undim selected buffer when focus is regained
Don’t check ‘auto-dim-other-buffers-dim-on-focus-out’ when focus is
regained.  Instead unconditionally undim selected buffer.  If the
variable is nil this will be a nop anyway (since selected buffer will
still be undimmed) but this new behaviour fixes a somewhat obscure
bug which stops the buffer from being undimmed when value of
‘auto-dim-other-buffers-dim-on-focus-out’ changes from t to nil while
Emacs’ frame is out of foucs.