Handle window and buffer switching at the same time
When a buffer is killed or buried at the same time as when window is
changed, the buffer shown in the previously selected window may remain
undimmed in Emacs prior to version 27.

For example, having two windows run ‘M-x describe-function RET defun
RET’ (which will show help in one of the windows), navigate to the
window showing *Help* buffer and quit the help by pressing ‘q’.  This
will select the other window and leave the new buffer in the remaining
window won’t be dimmed.

This happens because when we update state of buffers, we undim the
newly selected buffer and dim the old buffer.  Unfortunately, the old
buffer is no longer the one shown in the old window so that latter
operation doesn’t do anything useful.

Change the code to keep track of the last window (just like it keeps
track of last buffer) even outside of adow-mode.  With that new
information, whenever window changes, make sure that whatever is
displayed in the old window is dimmed.

Issue: https://github.com/mina86/auto-dim-other-buffers.el/issues/22