Fix buffer undimming when local variables are killed
The ‘kill-all-local-variables’ function kills all local variables and
face remapping.  This is a problem because adob is never notified of
the face remapping being removed.  If the buffer is not selected but
visible the result is that it becomes undimmed.

To reproduce the issue evaluate from a buffer using ‘eval-last-sexp’
(C-x C-e) or equivalent.  Do not use interactive ‘eval-expression’.

    (with-current-buffer (get-buffer "*adob-test*")
      (kill-all-local-variables))

Now set up at lest two windows one of which showing ‘*adob-test*’
buffer.  Select another window (‘*adob-test*’ should dim) and evaluate
the above again.  The test buffer should remain dimmed but becomes
undimmed even though it’s not the selected window.

There’s no hook that is called after the removal of face remapping and
marking ‘face-remapping-alist’ as a ‘permanent-local’ doesn’t sound
like a good idea so to fix the issue I’ve ended up adding an advice
around the ‘kill-all-local-variables’ function.