Allow configuring which faces to alter; not just the ‘default’
Introduce ‘auto-dim-other-buffers-affected-faces’ list which specifies
which faces should have relative remapping added and what that
remapping should be.  Previously, adob would only affect the ‘default’
face adding ‘auto-dim-other-buffers-face’ remapping to it.  With this
new variable, other faces can be configured to be changed as well.

For example, ‘org-block’ in Leuven theme uses light yellow background
which previously would stay yellow in dimmed windows.  Now, since
‘org-block’ is included in ‘auto-dim-other-buffers-affected-faces’,
it’s dimmed just like the rest of the window.

Note that no other faces are not affected by default so depending on
mode and colour theme various parts of the buffer may remain undimmed.
If that’s not desired, faces responsible for those colours need to be
manually added to the ‘auto-dim-other-buffers-affected-faces’ list.

Furthermore, since the list allows to set different remappings for
different faces, how each face is dimmed can be configured separately
as well by creating dimmed faces for each such face.

Fixes: https://github.com/mina86/auto-dim-other-buffers.el/issues/24