Introduce ‘auto-dim-other-buffers-hide-face’ for hidden dimmed text
Add ‘auto-dim-other-buffers-hide-face’ which has foreground and
background set to the same colour.  In effect, text rendered using
that face is invisible.

This is intended as a remapping for faces which hide the text such as
‘org-hide’ face (which otherwise show the text when window is dimmed).

The face is added to ‘auto-dim-other-buffers-affected-faces’ list as
a remapping for ‘org-hide’ but any other faces which are meant to hide
the text may need to be added separately to that list.

Fixes: https://github.com/mina86/auto-dim-other-buffers.el/issues/30