Add changing of the fringes in the selected window.
In addition to dimming non-selected windows, add fringe face to the
list of affected fringes highlighting it in the selected window.

This feature forces a more aggressive redrawing of the frames which
may cause flickering on systems which previously could handle
auto-dim-other-buffers without issues.  If this happens, remove the
fringe face from list of affected face by customising the
auto-dim-other-buffers-affected-faces variable or executing the
following:

    (setq auto-dim-other-buffers-affected-faces
          (assq-delete-all 'fringe auto-dim-other-buffers-affected-faces))

Fixes: https://github.com/mina86/auto-dim-other-buffers.el/issues/27